Add support for multiple scripting languages to the Script Engine
A Java utility (ScriptEngineUtility) is provided to share functionality between the scripts.
This functionality is written in Java, which essentially allows the scripts to execute Java code without having to actually be Java.

The arguments passed to and return results passed from the scripts are expected to be in JSON string format.
The ScriptEngineUtility should be used to convert to/from the JSON string to a JSON object for the script to process.

Two new exceptions are provided:
- ScriptEngineUnsupported = Error when the language (by file extension) is not available in the ScriptEngineType enum.
- ScriptEngineLoadFailed = Error when the engine does not exist (module is not added to the pom.xml or something went wrong while loading).

A pre-processor is added for handling the passed scripts.
This is necessary namely for Python, which must have the correct number of spaces/tabs prepended to it.
Currently, this is set to two spaces (both the engine.py and the pre-process method must have the same design).
Other languages can potentially utilize this pre-processor, but only Python currently gets pre-processed.

The engine.* resource files (such as engine.js) provide the engine template.
The util.* resource files (such as util.js) provide additional utilities that are language dependent and therefore are not available in the ScriptEngineUtility.

The following languages are supported:
- Javascript (.js)
- Java (.java)
- Python (.py)
- Groovy (.groovy)

The following languages could be supported, if problems are resolved:
- Ruby (.rb) (via JRuby)
- Perl (.pl) (via Rakudo)

Ruby is currently not working due to `java.lang.NoSuchFieldError: O_TMPFILE`.
More research is needed and an upstream bugreport is referenced below.

Perl could potentially be made to work via Rakudo, but is there a maven repository to add to the pom.xml?

The engine.rb and engine.pl are stubbed out and have not been tested.
Once the languages are imported and working properly, those engine.* files may need to be updated.

see: https://github.com/jruby/jruby/issues/5334
see: https://rakudo.org/