Sprint staging (#193)
* Count results and set count variable for file results (#190)

* Conditionally include attachment (#192)

* Conditionally include attachment

* Process include attachment as template

* Send email notification on workflow incident (#191)

* Send email notification on workflow incident

* Add configuration to specify environment

* Update raml json

* File task line count and read line (#194)

* File task line count and read line

Need to only count EOL but Java probably brings in whole line into memory

* Fix variable name

* Update generated JSON schema

* Check for existence of output variable (#196)

Only applicable to concrete class that extend output interface that do not require the output variable from the component persistent definition. In this case the AbstractDatabaseOutputDelegate implements Output but the DatabaseQueryTask does not require output variable for all types of its operations.