Issue 219: Remove shade plugin due to inherent problems and bugs.
Errors regarding multiple registrations of `AttributeConverter` have surfaced (in mod-workflow).
This is a bug between the shade plugin and hibernate as far as I can tell from my online research.

I believe that this is sufficient justification to deviate from the Folio practice of using the shade plugin.
Strip out the shade plugin entirely and update the `Dockerfile`.