MODCAMUNDA-1: BMP related changes, removing dead code and fixing log messages.
The bpm xml file appears to not be used anymore.
Remove the bpm xml file.

Add all of the bpm configuration settings to the yaml file with the defaults specified so that it is easier to figure them out.

The use of lombok for logging results in a static class name usage for the log.
This doesn't work well for an abstract class.
Instead, instantiate the log in a non-static way so that the implementation class name is used.

Remove the unused `camunda.spring.boot.version` from the `pom.xml` file.
Add a common version number setting for the graalvm.

Update the graalvm versions.