MODCAMUNDA-7: Add unit tests and fix sonarcloud code smells.
Switch to use `@Spy` instead of `@InjectMocks` in most cases to avoid problems where Mockito has conflicts when abstract-classes or interfaces are used multiple times with `@Mock`.
Add and use `@BeforeEach` in the case where `@InjectMocks` seems to be better served.

The delegate unit tests exposed a design bug where the abstract interfaces are calling a getter without checking if the value is NULL for the default abstract implementation.
There is a "has" function added to child abstract implementation that provides the expected safety check.
Move this child abstract class "has" method into the parent abstract class and update the default abstract getter method to call this "has" method.

This is intended to be changed into a PR into the base branch or the staging branch at the start of the next sprint.
The merge into master is only temporary and is intended only as a way to show this PR.

Re-design functional logic to address `S3776` reported by sonarcloud in some extreme cases.