De-serialized a stringified JSON code for the ScriptTask.
The `ScriptTask` should be serialized using `JSON.stringify()` in the Javascript.
Once this is done the server can now safely de-serialize the code.
The code no longer needs to have the newlines altered when added to the `code` property value in the fw-cli and similar clients.
Such code now requires the `JSON.stringify()` or something similar when adding it to the `code` property value.