Improve Module Descriptor and application.yml to favor auto-deployment.
Expose additional notable environment variables in the Module Descriptor.
The default settings provided, except for maybe the admin account name and associated passwords, should represent the sane default values for auto-deployment.

The ActiveMQ may be removed in the future but for now it is in use.
Expose the settings such that if a non-local ActiveMQ is used, then it can be appropriately configured.

The server port should not change from 8081 during deployment without additional module descriptor changes.
Document the server port setting as a practice.
The server context path is also exposed as a practice.

The Flyway should be disabled by default for auto-deployment (in general).

Expose the tenant information to make the required changes more obvious should a non-default tenant be used.

Expose notable Camunda BPM settings that may relate to or be influence by deployment.

The default port number is now set to 8081 rather than 9000.

Flyway is now disabled by default.

The Hibernate DDL auto is now set to `update` by default.