Fix NULL problem and logging problems.
The `f` can be NULL.
Handle that case such that the existing NULL check on the `value` handles it.

The `NONE` and `SIMPLE` types are known.
Don't call them "unknown".

The processor tasks warnings seem superfluous.
For now, just add another log for `Task` types that are not `ProcessorTask` types.
These logs may be good to entirely remove in the future.