Fix handling of the class name vs the delegate name.
I had not realized that the delegate name is the name text and not the class name.
This corrects the usage mistake to make the class name and delegate name distinct functions.

The misunderstanding stems from the `getExpression()` function that uses the class simple name and calls the resulting processed variable the `delegateName`.
This is now avoided by using a better named functions and variables.

Remove an unused function that I was going to use but ended up not using.