MODCAMUNDA-14: Improve the error logging regarding FileTask write problems.
The error described in the ticket relates to a lack of NULL checks while logging a warning.
Add a NULL check and make sure the log is printed.
Add an addition log specific to the NULL check and update the existing log.

This shows an underlying problem in either the crafted Workflow or some part of the process.